web/backend/requirements.txt
# ============================================================
# NEXUS TRADER Web — Backend Python Dependencies
# ============================================================

# FastAPI + ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis
redis[hiredis]>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
pydantic[email]>=2.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Rate Limiting
slowapi>=0.1.9
limits>=3.6.0

# Utilities
python-dotenv>=1.0.0
